Support additional LE options
Add method to support additional options via environment variable
LE_ADDITIONAL_OPTIONS.

The variable is presumed to be whitespace delimited but can also
contain quoted substrings which can be whitespace delimited, which
is a bit of a headache.

So far haven't figured out a way to make certbot absorb the
expanded/substituted command without complaint, so amending a hack-ish eval in the
interim.